src/retriever/requirements.txt
PyYAML==6.0.2
fastapi==0.115.13
pydantic==2.11.7
joblib==1.5.1
numpy==2.2.6
pandas==2.3.0
scikit-learn==1.7.0
uvicorn==0.27.0
faiss-cpu==1.11.0
cohere==5.15.0
